rename server pkg -> daemon
