+++ pkg/daemon/signaling/signaling.go
@@ -5,8 +5,8 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/trickstercache/trickster/v2/pkg/server/instance"
-	"github.com/trickstercache/trickster/v2/pkg/server/reload"
+	"github.com/trickstercache/trickster/v2/pkg/daemon/instance"
+	"github.com/trickstercache/trickster/v2/pkg/daemon/reload"
 )
 
 func Wait(si *instance.ServerInstance) {
