+++ pkg/proxy/handlers/reload.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/trickstercache/trickster/v2/pkg/cache"
 	"github.com/trickstercache/trickster/v2/pkg/config"
+	"github.com/trickstercache/trickster/v2/pkg/daemon/reload"
 	"github.com/trickstercache/trickster/v2/pkg/proxy/headers"
-	"github.com/trickstercache/trickster/v2/pkg/server/reload"
 )
 
 // ReloadHandleFunc will reload the running configuration if it has changed
